render-requirements.txt
fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.11.7
requests>=2.32.4
pymupdf>=1.26.3
python-docx>=1.2.0
google-genai>=1.28.0
pinecone-client>=7.3.0